SDI-50: 🔧 Initial stab at netty / webflux
