+++ src/main/kotlin/uk/gov/justice/digital/hmpps/nomisvisitsmappingservice/config/ResourceServerConfiguration.kt
@@ -3,25 +3,26 @@ package uk.gov.justice.digital.hmpps.nomisvisitsmappingservice.config
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.r2dbc.repository.config.EnableR2dbcRepositories
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity
+import org.springframework.security.config.annotation.method.configuration.EnableReactiveMethodSecurity
 import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity
 import org.springframework.security.config.web.server.ServerHttpSecurity
 import org.springframework.security.web.server.SecurityWebFilterChain
 
 @Configuration
 @EnableWebFluxSecurity
+@EnableReactiveMethodSecurity
 @EnableR2dbcRepositories
-@EnableGlobalMethodSecurity(prePostEnabled = true, proxyTargetClass = true)
 class ResourceServerConfiguration {
 
   @Bean
   fun springSecurityFilterChain(http: ServerHttpSecurity): SecurityWebFilterChain {
     return http
+      .csrf { it.disable() } // crst not needed an rest api
       .authorizeExchange {
         it.pathMatchers(
-          "/webjars/**", "/favicon.ico", "/csrf",
+          "/favicon.ico", "/csrf",
           "/health/**", "/info",
-          "/v3/api-docs/**", "/swagger-ui/**", "/swagger-ui.html"
+          "/v3/api-docs/**", "/webjars/swagger-ui/**", "/swagger-ui.html"
         ).permitAll()
           .anyExchange().authenticated()
       }
