+++ src/main/resources/application.yml
@@ -79,3 +79,7 @@ management:
       cache:
         time-to-live: 2000ms
 
+azure:
+  application-insights:
+    web:
+      enabled: false
