+++ src/test/kotlin/uk/gov/justice/digital/hmpps/nomisvisitsmappingservice/resource/OpenApiDocsTest.kt
@@ -21,7 +21,7 @@ class OpenApiDocsTest : IntegrationTestBase() {
   @Test
   fun `open api docs are available`() {
     webTestClient.get()
-      .uri("/swagger-ui/index.html?configUrl=/v3/api-docs")
+      .uri("/webjars/swagger-ui/index.html?configUrl=/v3/api-docs")
       .accept(MediaType.APPLICATION_JSON)
       .exchange()
       .expectStatus().isOk
